Restructure views in core & non-core
