+++ jolpica_api/f1_api/views/core/circuits.py
@@ -3,8 +3,8 @@ from drf_spectacular.utils import extend_schema, extend_schema_view
 from jolpica.formula_one import models as f1
 from jolpica_schemas.f1_api.alpha.core.circuit import CircuitQueryParams, CircuitSummary, RetrievedCircuitDetail
 
-from ..serializers import CircuitSerializer
-from ..utils import pydantic_to_open_api_parameters
+from ...serializers import CircuitSerializer
+from ...utils import pydantic_to_open_api_parameters
 from .base_viewset import BaseFilterableViewSet
